src/legal-mcp/requirements.txt
# Legal MCP Server Dependencies
# Core server dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.4.0

# MCP protocol dependencies
mcp>=1.0.0

# Database dependencies
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Vector and AI dependencies
sentence-transformers>=2.2.2
numpy>=1.24.0
pgvector>=0.2.0

# PDF and document processing
PyPDF2>=3.0.1

# Utility dependencies
python-dotenv>=1.0.0
aiofiles>=23.2.1

# CORS and middleware
python-multipart>=0.0.6